requirements.txt
asgiref==3.4.1
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.48.0
django-decouple==2.1
idna==3.3
mypy-extensions==0.4.3
oauthlib==3.2.0
pathspec==0.9.0
platformdirs==2.4.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
sqlparse==0.4.2
tomli==2.0.0
typing_extensions==4.0.1
urllib3==1.26.7
